requirements.txt
ansible==2.2.1.0
appdirs==1.4.2
boto==2.46.1
cffi==1.9.1
cryptography==1.7.2
enum34==1.1.6
idna==2.4
ipaddress==1.0.18
Jinja2==2.8.1
MarkupSafe==0.23
ntlm-auth==1.0.2
ordereddict==1.1
packaging==16.8
paramiko==2.1.2
pyasn1==0.2.3
pycparser==2.17
pycrypto==2.6.1
pyparsing==2.1.10
pywinrm==0.2.2
PyYAML==3.12
requests==2.13.0
requests-ntlm==1.0.0
six==1.10.0
xmltodict==0.10.2
